PHP Practice
I made a simple form and tested it to see if it could write to a .txt
file.  It was successful on the server Dr. Farzan gave us.  Try it on
your server, and read that 'test.txt' file that will be created from it.
It will allow multiple entries, appending to the end of the file.  I
believe this is the start of what we want.