Added PHP to add new customer to database
I attempted to create a PHP file that would add all customer information
into the appropriate areas of the database, following instructions
provided in 'About_Sample_Data'.  I did not test it, and some changes
may need to be made for it to work properly, but it should be a good
start.

Will need to fix line 50($link) to work.